requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.11.5
cryptography==2.4.2
idna==2.8
Jinja2==2.10
lxml==4.3.0
MarkupSafe==1.1.0
ncclient==0.6.3
paramiko==2.4.2
pyaml==18.11.0
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
PyYAML==3.13
selectors2==2.0.1
six==1.12.0
xmltodict==0.11.0
